+++ README-template.md
@@ -5,10 +5,11 @@
 [![Custom badge](https://img.shields.io/endpoint?url=https%3A%2F%2Fdeno-visualizer.danopia.net%2Fshields%2Fcache-size%2Fhttps%2Fdeno.land%2Fx%2Flapi%2Fmod.ts)](https://deno.land/x/lapi)
 [![Custom badge](https://img.shields.io/endpoint?url=https%3A%2F%2Fdeno-visualizer.danopia.net%2Fshields%2Fdep-count%2Fhttps%2Fdeno.land%2Fx%2Flapi%2Fmod.ts)](https://deno.land/x/lapi)
 [![Custom badge](https://img.shields.io/endpoint?url=https%3A%2F%2Fdeno-visualizer.danopia.net%2Fshields%2Fupdates%2Fhttps%2Fdeno.land%2Fx%2Flapi%2Fmod.ts)](https://deno.land/x/lapi)
+[![deno doc](https://doc.deno.land/badge.svg)](https://doc.deno.land/https/deno.land/x/lapi/mod.ts)
 
 A full featured middleware based web framework for Deno.
 
-Roadmap: https://github.com/LukeShay/lapi/discussions/28
+[Checkout our roadmap here!](https://github.com/LukeShay/lapi/discussions/28)
 
 **NOTE:** This project is a heavy WIP, therefore this may not be up to date.
 
@@ -110,6 +111,25 @@ router.use("GET", "/", (ctx) => {
 application.use(router.routes()).start();
 ```
 
+### Path Parameters
+
+The routing middleware provides path parameters as will. Under the hood, we are
+using [pillarjs/path-to-regexp](https://github.com/pillarjs/path-to-regexp).
+
+```typescript
+import { Application } from "https://deno.land/x/lapi/mod.ts";
+import { Router } from "https://deno.land/x/lapi/middleware/router.ts";
+
+const application = new Application();
+const router = new Router();
+
+router.use("GET", "/:name", (ctx) => {
+  ctx.response.body = `Hello, ${ctx.request.pathParams.name}!`;
+});
+
+application.use(router.routes()).start();
+```
+
 ## Cors
 
 There is a Middleware provided in `middleware/cors.ts` that will set up the
@@ -129,6 +149,46 @@ application.use(cors()).use((ctx) => {
 });
 
 application.start();
+```
+
+## Native HTTP Server
+
+Lapi is capable of using the standard and native http server. By default, Lapi
+will use the standard http server, but if you run with the `--unstable` flag, it
+will use the native http server. You can force Lapi to use either one of these.
+
+> note: you must run with `--unstable` for the native http server to start.
+
+Force the standard http server:
+
+```typescript
+import { Application, HttpServerStd } from "https://deno.land/x/lapi/mod.ts";
+
+const application = new Application({ server: new HttpServerStd() });
+
+application.use((ctx) => {
+  if (ctx.request.method === "GET") {
+    ctx.response.body = "Hello";
+  }
+});
+
+await application.start();
+```
+
+Force the native http server:
+
+```typescript
+import { Application, HttpServerNative } from "https://deno.land/x/lapi/mod.ts";
+
+const application = new Application({ server: new HttpServerNative() });
+
+application.use((ctx) => {
+  if (ctx.request.method === "GET") {
+    ctx.response.body = "Hello";
+  }
+});
+
+await application.start();
 ```
 
 # Roadmap
