+++ application.ts
@@ -3,12 +3,29 @@
 
 import { compose, ComposedMiddleware, Middleware } from "./middleware.ts";
 import { HttpServer } from "./http_server.ts";
+import { HttpServerNative } from "./http_server_native.ts";
 import { HttpServerStd } from "./http_server_std.ts";
+import { isUnstable } from "./utils.ts";
 import { updateTypeAndGetBody } from "./response.ts";
 
+/** Options for configuring an Application. */
 export interface ApplicationOptions<T> {
+  /**
+   * Specifies the port to run on.
+   *
+   * @default 3000
+   */
   port?: number;
+
+  /** Specifies the host to listen to. */
   host?: string;
+
+  /**
+   * Selects the the type of HTTP server you want to use. If you want to run the
+   * native HTTP server, you must use the '--unstable' flag. Lapi determines
+   * this by see if it has been run with '--unstable'. If it has, the native
+   * HTTP server will be used. Otherwise, the standard HTTP server is used.
+   */
   server?: HttpServer<T>;
 }
 
@@ -23,8 +40,12 @@ export class Application<T> {
 
   /** Constructs an Application. */
   constructor({ port, host, server }: ApplicationOptions<T> = {}) {
-    this.#httpServer = server ||
-      (new HttpServerStd({
+    this.#httpServer = server || isUnstable()
+      ? (new HttpServerNative({
+        port: port || 3000,
+        host,
+      }) as unknown as HttpServer<T>)
+      : (new HttpServerStd({
         port: port || 3000,
         host,
       }) as unknown as HttpServer<T>);
