+++ application_test.ts
@@ -1,3 +1,5 @@
+// Copyright 2020 Luke Shay. All rights reserved. MIT license.
+
 import { assert } from "./deps_test.ts";
 import { Application } from "./application.ts";
