+++ deps_test.ts
@@ -1,1 +1,3 @@
+// Copyright 2020 Luke Shay. All rights reserved. MIT license.
+
 export * from "https://deno.land/std@0.99.0/testing/asserts.ts";
