+++ mod.ts
@@ -2,15 +2,16 @@
 /* @module lapi */
 
 export * from "./application.ts";
-export * from "./application.ts";
 export * from "./context.ts";
+export * from "./http_server_native.ts";
+export * from "./http_server_std.ts";
 export * from "./http_server.ts";
 export * from "./middleware.ts";
+export * from "./renderer_native.ts";
+export * from "./renderer_std.ts";
 export * from "./renderer.ts";
+export * from "./request_native.ts";
+export * from "./request_std.ts";
 export * from "./request.ts";
 export * from "./response.ts";
-export * from "./http_server_std.ts";
-export * from "./request_std.ts";
-export * from "./http_server_native.ts";
-export * from "./request_native.ts";
 export type { Status } from "./deps.ts";
