+++ renderer.ts
@@ -1,5 +1,7 @@
-import { Body, BodyFunction } from "./response.ts";
-import { convertBodyToBodyInit, convertBodyToStdBody } from "./oak.ts";
+// Copyright 2020 Luke Shay. All rights reserved. MIT license.
+/* @module lapi/renderer */
+
+import { Body } from "./response.ts";
 
 export interface Rendered<T> {
   body?: T;
@@ -8,19 +10,4 @@ export interface Rendered<T> {
 
 export interface Renderer<T> {
   (body: Body, type?: string | null): Promise<Rendered<T>> | Rendered<T>;
-}
-
-export async function defaultStdRenderer(
-  body: Body | BodyFunction,
-  type?: string | null,
-): Promise<Rendered<Uint8Array | Deno.Reader>> {
-  const [resultBody, resultType] = await convertBodyToStdBody(body, type);
-  return { body: resultBody, type: resultType };
-}
-export async function defaultNativeRenderer(
-  body: Body | BodyFunction,
-  type?: string | null,
-): Promise<Rendered<BodyInit>> {
-  const [resultBody, resultType] = await convertBodyToBodyInit(body, type);
-  return { body: resultBody, type: resultType };
 }
